ci: Add a call to the update label backport action
Add an action to call the workflow that update the labels of backported
PRs in v1.14 stable branch.

This commit is based on the following commits by Fabio from v1.14 branch:
- 81ade5f693b8 ("ci: Call the workflow to update labels of backported PRs")
- a5a047f2fa84 ("ci: Use pull_request_target in update label workflow")

The primary change here is to list all maintained branches in a single
workflow on main in order to simplify the maintenance burden when
creating new stable branches (eg, during v1.15 stable branch creation).

Signed-off-by: Joe Stringer <joe@cilium.io>